+++ tests/plugins/find/find.js
@@ -146,5 +146,223 @@ bender.test( {
 
 			dialog.getButton( 'cancel' ).click();
 		} );
+	},
+
+	// (#4987)
+	'test find text with double space between words': function() {
+		var bot = this.editorBot;
+
+		bot.setHtmlWithSelection( '<p>example&nbsp; text</p>' );
+
+		bot.dialog( 'find', function( dialog ) {
+			dialog.setValueOf( 'find', 'txtFindFind', 'example  text' );
+			dialog.getContentElement( 'find', 'btnFind' ).click();
+
+
+			assert.areSame( '<p><span title="highlight">example&nbsp; text</span></p>', bot.getData( true ) );
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test find text with double &nbsp; between words': function() {
+		var bot = this.editorBot;
+
+		bot.setHtmlWithSelection( '<p>example&nbsp;&nbsp;text</p>' );
+
+		bot.dialog( 'find', function( dialog ) {
+			dialog.setValueOf( 'find', 'txtFindFind', 'example  text' );
+			dialog.getContentElement( 'find', 'btnFind' ).click();
+
+
+			assert.areSame( '<p><span title="highlight">example&nbsp;&nbsp;text</span></p>', bot.getData( true ) );
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test find text with double space between words in read-only mode': function() {
+		var bot = this.editorBot;
+
+		bot.setHtmlWithSelection( '<p>example&nbsp; text</p>' );
+		bot.editor.setReadOnly( true );
+
+		bot.dialog( 'find', function( dialog ) {
+			dialog.setValueOf( 'find', 'txtFindFind', 'example  text' );
+			dialog.getContentElement( 'find', 'btnFind' ).click();
+
+			bot.editor.setReadOnly( false );
+
+			assert.areSame( '<p><span title="highlight">example&nbsp; text</span></p>', bot.getData( true ) );
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test find word when pattern starting with empty space': function() {
+		var bot = this.editorBot;
+
+		bot.setHtmlWithSelection( '<p>example&nbsp;[&nbsp;text]</p>' );
+
+		bot.dialog( 'find', function( dialog ) {
+			dialog.getContentElement( 'find', 'btnFind' ).click();
+
+			assert.areSame( '<p>example&nbsp;<span title="highlight">&nbsp;text</span></p>', bot.getData( true ) );
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test find text without spaces between words should thrown alert with proper message': function() {
+		var bot = this.editorBot,
+			spy = sinon.stub( window, 'alert' );
+
+		bot.setHtmlWithSelection( '<p>exampletext</p>' );
+
+		bot.dialog( 'find', function( dialog ) {
+			dialog.setValueOf( 'find', 'txtFindFind', 'example  text' );
+			dialog.getContentElement( 'find', 'btnFind' ).click();
+
+			assert.isTrue( spy.calledOnce, 'Find text without spaces between words should thrown alert' );
+			assert.areEqual( spy.args[ 0 ][ 0 ], this.editorBot.editor.lang.find.notFoundMsg,
+				'Find text without spaces between words should have proper alert message' );
+
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test find and replace text with double space between words': function() {
+		var bot = this.editorBot;
+
+		bot.setHtmlWithSelection( '<p>example&nbsp; text from CKEditor4</p>' );
+
+		bot.dialog( 'replace', function( dialog ) {
+			dialog.setValueOf( 'replace', 'txtFindReplace', 'example  text' );
+			dialog.setValueOf( 'replace', 'txtReplace', 'changed example text' );
+			dialog.getContentElement( 'replace', 'btnFindReplace' ).click();
+			dialog.getContentElement( 'replace', 'btnFindReplace' ).click();
+
+			assert.areSame( '<p><span title="highlight">changed example text</span> from ckeditor4</p>', bot.getData( true ) );
+
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test replace all text with double spaces between words': function() {
+		var bot = this.editorBot;
+
+		bot.setHtmlWithSelection( '<p>[example&nbsp; text]</p><p>example&nbsp; text</p>' );
+
+		bot.dialog( 'replace', function( dialog ) {
+			dialog.setValueOf( 'replace', 'txtReplace', 'replaced text' );
+			dialog.getContentElement( 'replace', 'btnReplaceAll' ).click();
+			dialog.getButton( 'cancel' ).click();
+
+			assert.areSame( '<p>replaced text</p><p>replaced text</p>', bot.getData( false, true ) );
+
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test space separator: SPACE': function() {
+		var bot = this.editorBot,
+			expected = '<p>test<span title="highlight">&nbsp; </span>test</p>';
+
+		bot.setHtmlWithSelection( '<p>test&nbsp; test</p>' );
+
+		bot.dialog( 'find', function( dialog ) {
+			dialog.setValueOf( 'find', 'txtFindFind', '  ' );
+			dialog.getContentElement( 'find', 'btnFind' ).click();
+
+			assert.areSame( expected, bot.getData( true ), 'Word separator SPACE is incorrect' );
+			dialog.getButton( 'cancel' ).click();
+		} );
+	},
+
+	// (#4987)
+	'test space separator: OGHAM SPACE MARK': function() {
+		assertSpaceSeparator( this.editorBot, '\u1680', 'OGHAM SPACE MARK' );
+	},
+
+	// (#4987)
+	'test space separator: EN QUAD': function() {
+		assertSpaceSeparator( this.editorBot, '\u2000', 'EN QUAD' );
+	},
+
+	// (#4987)
+	'test space separator: EM QUAD': function() {
+		assertSpaceSeparator( this.editorBot, '\u2001', 'EM QUAD' );
+	},
+
+	// (#4987)
+	'test space separator: EN SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2002', 'EN SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: EM SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2003', 'EM SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: THREE-PER-EM SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2004', 'THREE-PER-EM SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: FOUR-PER-EM SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2005', 'FOUR-PER-EM SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: SIX-PER-EM SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2006', 'SIX-PER-EM SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: FIGURE SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2007', 'FIGURE SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: PUNCTUATION SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2008', 'PUNCTUATION SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: THIN SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u2009', 'THIN SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: HAIR SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u200A', 'HAIR SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: NARROW NO-BREAK SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u202F', 'NARROW NO-BREAK SPACE' );
+	},
+
+	// (#4987)
+	'test space separator: IDEOGRAPHIC SPACE': function() {
+		assertSpaceSeparator( this.editorBot, '\u3000', 'IDEOGRAPHIC SPACE' );
 	}
+
 } );
+
+function assertSpaceSeparator( bot, unicode, name ) {
+	var expected = '<p>test<span title="highlight">' + unicode + ' </span>test</p>';
+
+	bot.setHtmlWithSelection( '<p>test[' + unicode + ' ]test</p>' );
+
+	bot.dialog( 'find', function( dialog ) {
+		dialog.getContentElement( 'find', 'btnFind' ).click();
+
+		assert.areSame( expected, bot.getData( true ), 'Word separator ' + name + ' is incorrect' );
+		dialog.getButton( 'cancel' ).click();
+	} );
+}